feat(global_storage): Move `entity` module in `kv_store`
