--- src/service/src/features/global_storage/entity.rs
@@ -1,21 +1,0 @@
-// prose-pod-api
-//
-// Copyright: 2025, Rémi Bardon <remi@remibardon.name>
-// License: Mozilla Public License v2.0 (MPL v2.0)
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "kv_store")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub namespace: String,
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub key: String,
-    pub value: Json,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {}
-
-impl ActiveModelBehavior for ActiveModel {}
