+++ src/service/src/features/global_storage/mod.rs
@@ -3,7 +3,6 @@
 // Copyright: 2025, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
-mod entity;
 mod kv_store;
 pub(crate) mod migrations;
